zcash_pool_migration: extract the wallet trait seams into `backend`
The five traits a wallet implements to become a migration wallet were spread
through `engine.rs`: the planning and store seams near the top, the crypto
and proving seams six hundred lines further down, interleaved with the
operations that consume them. Reading the interface meant reading the
driver.

Move all five to `backend`, in one screen, ordered by what each costs a
wallet to provide: `MigrationBackend` and `PoolMigrationRead` /
`PoolMigrationWrite` need no key material, `MigrationCrypto` needs the
account's viewing key and spend authority, and `MigrationProver` needs
mutable commitment-tree access and the proving parameters. That ordering is
the reason the split exists at all, so it is now visible.

A pure move. The edits are doc links that now cross a module boundary, plus
one that pointed at the engine's private `Signing` enum and so could never
have resolved from a public page; it is stated in prose instead. All five
are re-exported from `engine`, so existing paths keep resolving.
